Refactor World Zone
