+++ package.json
@@ -27,7 +27,7 @@
 		"react-countdown": "^2.3.2",
 		"react-dom": "^17.0.2",
 		"react-hooks-visible": "^1.1.1",
-		"react-scripts": "5.0.1",
+		"react-scripts": "^2.1.3",
 		"signals": "^1.0.0",
 		"split-grid": "^1.0.11",
 		"three": "^0.139.2",
